Including changes for setting up linters and basic fastapi code
In this commit i have tried to setup the poetry with pyproject.toml
poetry seems to be nice way to manage dependencies but this is
conflicting with toml as tox is not using poetry installed dependencies

So I have gone with poetry which seems simple. For running linters
I have created a custom script which is included in pyproject.toml
which can be trigger with command poetry run lint.